Move fee settings text to its own components
This was part of the fee settings button, but in certain cases we want
the exact same text without the button—in particular, swaps have a
different way of bringing up transaction/network settings, but still
need to display the current fee settings at the top level.
